+++ src/main/java/org/opensearch/sdk/ExtensionsRunner.java
@@ -18,15 +18,9 @@ import org.apache.logging.log4j.Logger;
 import org.opensearch.Version;
 import org.opensearch.cluster.ClusterModule;
 import org.opensearch.cluster.node.DiscoveryNode;
-import org.opensearch.common.io.stream.InputStreamStreamInput;
-import org.opensearch.common.io.stream.NamedWriteable;
-import org.opensearch.common.io.stream.NamedWriteableAwareStreamInput;
 import org.opensearch.common.io.stream.NamedWriteableRegistry;
 import org.opensearch.common.io.stream.NamedWriteableRegistryParseRequest;
-import org.opensearch.common.io.stream.NamedWriteableRegistryParseResponse;
-import org.opensearch.extensions.DefaultExtensionPointRequest;
-import org.opensearch.common.io.stream.NamedWriteableRegistryResponse;
-import org.opensearch.common.io.stream.StreamInput;
+import org.opensearch.extensions.OpenSearchRequest;
 import org.opensearch.common.network.NetworkModule;
 import org.opensearch.common.network.NetworkService;
 import org.opensearch.common.settings.Settings;
@@ -48,19 +42,13 @@ import org.opensearch.transport.ConnectionManager;
 import org.opensearch.transport.TransportService;
 import org.opensearch.transport.TransportSettings;
 import org.opensearch.transport.TransportInterceptor;
-
+import org.opensearch.transport.TransportResponse;
 import org.opensearch.sdk.netty4.Netty4Transport;
 import org.opensearch.sdk.netty4.SharedGroupFactory;
-
-import java.io.ByteArrayInputStream;
 import java.io.File;
 import java.io.IOException;
-import java.io.InputStream;
-import java.util.ArrayList;
 import java.util.Collections;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 import java.util.function.Function;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
@@ -71,11 +59,6 @@ import static org.opensearch.common.UUIDs.randomBase64UUID;
 public class ExtensionsRunner {
     private ExtensionSettings extensionSettings = readExtensionSettings();
     private DiscoveryNode opensearchNode;
-    private List<NamedWriteableRegistry.Entry> namedWriteables = getNamedWriteables();
-    final private NamedWriteableRegistry namedWriteableRegistry = new NamedWriteableRegistry(namedWriteables);
-
-    public ExtensionsRunner() throws IOException {}
-
     private final Settings settings = Settings.builder()
         .put("node.name", extensionSettings.getExtensionname())
         .put(TransportSettings.BIND_HOST.getKey(), extensionSettings.getHostaddress())
@@ -84,20 +67,15 @@ public class ExtensionsRunner {
     private final Logger logger = LogManager.getLogger(ExtensionsRunner.class);
     private final TransportInterceptor NOOP_TRANSPORT_INTERCEPTOR = new TransportInterceptor() {
     };
+    private NamedWriteableRegistryApi namedWriteableRegistryApi = new NamedWriteableRegistryApi();
+
+    public ExtensionsRunner() throws IOException {}
 
     private ExtensionSettings readExtensionSettings() throws IOException {
         File file = new File(ExtensionSettings.EXTENSION_DESCRIPTOR);
         ObjectMapper objectMapper = new ObjectMapper(new YAMLFactory());
         ExtensionSettings extensionSettings = objectMapper.readValue(file, ExtensionSettings.class);
         return extensionSettings;
-    }
-
-    // current placeholder for extension point override getNamedWriteables()
-    private List<NamedWriteableRegistry.Entry> getNamedWriteables() {
-
-        List<NamedWriteableRegistry.Entry> namedWriteables = new ArrayList<>();
-        // extensions will add named writeable registry entries here
-        return namedWriteables;
     }
 
     private void setOpensearchNode(DiscoveryNode opensearchNode) {
@@ -116,60 +94,21 @@ public class ExtensionsRunner {
         return pluginResponse;
     }
 
-    NamedWriteableRegistryResponse handleNamedWriteableRegistryRequest(DefaultExtensionPointRequest request) {
-
-        logger.info("Registering Named Writeable Registry Request recieved from OpenSearch.");
-
-        // iterate through Extensions's named writeables and add to extension entries
-        Map<String, Class> extensionEntries = new HashMap<>();
-        for (NamedWriteableRegistry.Entry entry : this.namedWriteables) {
-            extensionEntries.put(entry.name, entry.categoryClass);
+    /**
+     * Handles a request from OpenSearch and invokes the extension point API corresponding with the request type
+     *
+     * @param request  The request to handle.
+     * @return A response to OpenSearch for the corresponding API
+     */
+    TransportResponse handleOpenSearchRequest(OpenSearchRequest request) {
+        // Read enum
+        switch (request.getRequestType()) {
+            case REQUEST_OPENSEARCH_NAMED_WRITEABLE_REGISTRY:
+                return namedWriteableRegistryApi.handleNamedWriteableRegistryRequest(request);
+            // Add additional request handlers here
+            default:
+                return null;
         }
-        NamedWriteableRegistryResponse namedWriteableRegistryResponse = new NamedWriteableRegistryResponse(extensionEntries);
-        return namedWriteableRegistryResponse;
-    }
-
-    <C extends NamedWriteable> NamedWriteableRegistryParseResponse handleNamedWriteableRegistryParseRequest(
-        NamedWriteableRegistryParseRequest request
-    ) throws IOException {
-
-        logger.info("Registering Named Writeable Registry Parse request from OpenSearch");
-        boolean status = false;
-
-        // extract data from request and procress fully qualified category class name into class instance
-        Class<C> categoryClass = (Class<C>) request.getCategoryClass();
-        byte[] context = request.getContext();
-
-        // transform byte array context into an input stream
-        try (InputStream inputStream = new ByteArrayInputStream(context, 0, context.length)) {
-
-            // convert input stream to stream input
-            try (
-                StreamInput streamInput = new NamedWriteableAwareStreamInput(
-                    new InputStreamStreamInput(inputStream),
-                    namedWriteableRegistry
-                )
-            ) {
-
-                // NamedWriteableAwareStreamInput extracts name from StreamInput, then uses both category class and name to extract
-                // reader from provided registry
-                // reader is then applied to the StreamInput object generated from the byte array (context)
-                try {
-                    C c = streamInput.readNamedWriteable(categoryClass);
-
-                    // TODO : current parse response to OpenSearch includes only the status of the parse request. Further research
-                    // needed to determine the workflow for extensions to utilize parsed objects after deserialization but this will be
-                    // within the scope of dynamic registration.
-                    status = true;
-                } catch (UnsupportedOperationException e) {
-                    logger.info("Failed to parse named writeable", e);
-                }
-
-            }
-        }
-
-        NamedWriteableRegistryParseResponse namedWriteableRegistryParseResponse = new NamedWriteableRegistryParseResponse(status);
-        return namedWriteableRegistryParseResponse;
     }
 
     IndicesModuleResponse handleIndicesModuleRequest(IndicesModuleRequest indicesModuleRequest, TransportService transportService) {
@@ -267,19 +206,19 @@ public class ExtensionsRunner {
         );
 
         transportService.registerRequestHandler(
-            ExtensionsOrchestrator.REQUEST_EXTENSION_NAMED_WRITEABLE_REGISTRY,
+            ExtensionsOrchestrator.REQUEST_OPENSEARCH_NAMED_WRITEABLE_REGISTRY,
             ThreadPool.Names.GENERIC,
             false,
             false,
-            DefaultExtensionPointRequest::new,
-            (request, channel, task) -> channel.sendResponse(handleNamedWriteableRegistryRequest(request))
+            OpenSearchRequest::new,
+            (request, channel, task) -> channel.sendResponse(handleOpenSearchRequest(request))
         );
 
         transportService.registerRequestHandler(
-            ExtensionsOrchestrator.REQUEST_EXTENSION_PARSE_NAMED_WRITEABLE,
+            ExtensionsOrchestrator.REQUEST_OPENSEARCH_PARSE_NAMED_WRITEABLE,
             ThreadPool.Names.GENERIC,
             NamedWriteableRegistryParseRequest::new,
-            (request, channel, task) -> channel.sendResponse(handleNamedWriteableRegistryParseRequest(request))
+            (request, channel, task) -> channel.sendResponse(namedWriteableRegistryApi.handleNamedWriteableRegistryParseRequest(request))
         );
 
         transportService.registerRequestHandler(
